+++ src/android/tools/FileProvider.java
@@ -1,0 +1,3 @@
+package de.kolbasa.apkupdater.tools;
+
+public class FileProvider extends androidx.core.content.FileProvider {}