+++ src/ionic/declarations/interfaces/App.d.ts
@@ -1,0 +1,22 @@
+declare module 'cordova-plugin-apkupdater' {
+
+    interface App {
+
+        /**
+         * App name.
+         */
+        name: string;
+
+        /**
+         * Package name.
+         */
+        package: string;
+
+        /**
+         * App version.
+         */
+        version: Version;
+
+    }
+
+}