+++ src/ionic/declarations/interfaces/AuthConfig.d.ts
@@ -1,0 +1,11 @@
+declare module 'cordova-plugin-apkupdater' {
+
+    interface AuthConfig {
+
+        user: string;
+
+        password: string;
+
+    }
+
+}