+++ src/ionic/declarations/interfaces/Progress.d.ts
@@ -1,0 +1,16 @@
+declare module 'cordova-plugin-apkupdater' {
+
+    interface Progress {
+
+        /**
+         * Flaoting point.
+         */
+        progress: number;
+
+        bytes: number;
+
+        bytesWritten: number;
+
+    }
+
+}