+++ src/ionic/declarations/interfaces/Version.d.ts
@@ -1,0 +1,17 @@
+declare module 'cordova-plugin-apkupdater' {
+
+    interface Version {
+
+        /**
+         * For example: "10000".
+         */
+        code: number;
+
+        /**
+         * For example: "1.0.0".
+         */
+        name: string;
+
+    }
+
+}