+++ www/ApkUpdater.js
@@ -21,7 +21,7 @@ module.exports = {
      * @param {object | undefined} opt - Optional
      * @param {string=} opt.zipPassword
      * @param {object=} opt.basicAuth
-     * @param {string=} opt.basicAuth.userId
+     * @param {string=} opt.basicAuth.user
      * @param {string=} opt.basicAuth.password
      * @param {function({progress: number, bytes: number, bytesWritten: number}): void=} opt.onDownloadProgress
      * @param {function({progress: number, bytes: number, bytesWritten: number}): void=} opt.onUnzipProgress